# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для dashbord'а эффективности сотрудников.",
        "title": "Employee Dashboard API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/dashboard/summary": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает показатели сотрудников (admin/manager — все, employee — свои)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Получение сводных показателей эффективности",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Начало периода",
                        "name": "date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конец периода",
                        "name": "date_to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.DashboardSummaryResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "Аутентификация пользователя и выдача JWT-токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход пользователя",
                "parameters": [
                    {
                        "description": "Login payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.DashboardSummaryResponse": {
            "type": "object",
            "properties": {
                "avgLoad": {
                    "type": "number"
                },
                "avgProductivity": {
                    "type": "number"
                },
                "deptEfficiency": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.DeptEfficiencyItem"
                    }
                },
                "monthlyStats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.MonthlyStat"
                    }
                },
                "overtime": {
                    "type": "number"
                },
                "satisfaction": {
                    "type": "number"
                },
                "topEfficiency": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.TopEfficiencyItem"
                    }
                },
                "topOvertime": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.TopOvertimeItem"
                    }
                }
            }
        },
        "controllers.DeptEfficiencyItem": {
            "type": "object",
            "properties": {
                "avg_productivity": {
                    "type": "number"
                },
                "department": {
                    "type": "string"
                },
                "job_level": {
                    "type": "string"
                }
            }
        },
        "controllers.LoginRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "controllers.MonthlyStat": {
            "type": "object",
            "properties": {
                "load": {
                    "type": "number"
                },
                "month": {
                    "type": "string"
                },
                "overtime": {
                    "type": "number"
                }
            }
        },
        "controllers.TopEfficiencyItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "overtime": {
                    "type": "number"
                },
                "productivity": {
                    "type": "number"
                },
                "tasks": {
                    "type": "integer"
                }
            }
        },
        "controllers.TopOvertimeItem": {
            "type": "object",
            "properties": {
                "job_satisfaction": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "overtime_hours": {
                    "type": "number"
                },
                "tasks_completed_per_day": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}